requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
bcrypt==4.0.1
cachelib==0.10.2
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
filelock==3.10.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
platformdirs==3.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
virtualenv==20.21.0
Werkzeug==2.2.3
WTForms==3.0.1
